Added /dist to gitignore and cleaned up node script
Made the package.json update code an IIFE instead of a named function to be cleaner

Added /dist to .gitignore to exclude the current Webpack build process. I didn't change the location to /build in the config because of the different structure. I may look deeper into the settings create-react-app uses later and replicate it, but for now this will do.